refactor(router): move rule tests into their own folder
