+++ modules/angular2/test/router/rules/route_paths/param_route_path_spec.ts
@@ -10,8 +10,8 @@ import {
   SpyObject
 } from 'angular2/testing_internal';
 
-import {ParamRoutePath} from 'angular2/src/router/rules/param_route_path';
-import {parser, Url} from 'angular2/src/router/rules/url_parser';
+import {ParamRoutePath} from 'angular2/src/router/rules/route_paths/param_route_path';
+import {parser, Url} from 'angular2/src/router/url_parser';
 
 export function main() {
   describe('PathRecognizer', () => {
